requirements.txt
suds-jurko==0.6
urllib3==1.23